requirements.txt
Django==2.2.1
django-haystack==2.8.1
django-tinymce==2.8.0
jieba==0.39
Pillow==6.0.0
pytz==2019.1
sqlparse==0.3.0
Whoosh==2.7.4
